Refactor lastDelimIndex counter
The loop in ParserMarkdown tries to find if there's metadata in the file and locate the last "+++" delimiter.
But, I overlooked that each line ends with "\n", causing the issue of hard-coding [lastDelimIndex+4:] to get
a new slice without metadata. Fixing this lets us use [lastDelimIndex:],
since the counter now considers the newline character.